Add Sun Intensity Curve
Reduces the sun's light beyond Eeloo's orbit. The lighting should be nearly identical to stock, before slowly dimming at the edge of the Sun's SOI. Change brightnessCurve to dim sunlight to a point.

These changes are necessary for all stars in the Interstellar Consortium, or all stars would be shining and everything would be too bright. IC-plugin should set a good example for the bare minimum.